+++ index.html
@@ -102,6 +102,7 @@
     </footer>
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.18.1/moment.min.js"></script>
-    <script src="js/script.js"></script>
+    <script type="module" src="js/data.js"></script>
+    <script type="module" src="js/script.js"></script>
   </body>
 </html>
